+++ MarkdownMonster/Windows/Helpers/ValueConverters.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Windows;
 using System.Windows.Data;
 using System.Windows.Media.Imaging;
+using LibGit2Sharp;
 using Westwind.Utilities;
 
 namespace MarkdownMonster.Windows
@@ -242,7 +243,38 @@ namespace MarkdownMonster.Windows
         }
     }
 
-    
+    internal class SourceControlIconConverter : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            if (!(value is FileStatus))
+                return value;
+
+            var status = (FileStatus) value;
+
+            if (status == FileStatus.Unaltered)
+                return SourceControlIcons.Normal;
+            if (status == FileStatus.ModifiedInIndex || status == FileStatus.ModifiedInWorkdir) 
+                return SourceControlIcons.Changed;
+            
+            if (status == FileStatus.Ignored)
+                return SourceControlIcons.Ignored;
+            if (status == FileStatus.NewInIndex || status == FileStatus.NewInWorkdir)
+                return SourceControlIcons.Added;
+            if (status == FileStatus.Conflicted)
+                return SourceControlIcons.Conflict;
+
+            return null;
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            throw new NotImplementedException();
+        }
+    }
+
+
+
     public class UriToCachedImageConverter : IValueConverter
     {
         private static Dictionary<string, BitmapImage> CachedBitmapImages = new Dictionary<string, BitmapImage>();