+++ Changelog.md
@@ -6,7 +6,10 @@
 [![Chocolatey](https://img.shields.io/chocolatey/dt/markdownmonster.svg)](https://chocolatey.org/packages/MarkdownMonster)
 [![Web Site](https://img.shields.io/badge/Markdown_Monster-WebSite-blue.svg)](https://markdownmonster.west-wind.com)
 
-### 1.10.10
+### 1.10.12
+
+* **Git File Status in File Browser**  
+The Folder browser now shows Git status information via icons. There are also a new option to undo changes on the file context menu.
 
 * **Add Clone Git Repository**  
 Added option to clone a Git repository to a local folder to make it easier to retrieve Git content for local editing. Also added a new **File -> Git** menu that houses this option and **Commit to Git** (same behavior as the Tab context menu). **Requires:** that Git and Git Credential Manager (for private or authenticated repos) are installed.